pyproject.toml
[project]
name = "artlangvocalsynth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "japanize-matplotlib>=1.1.3",
    "music21>=9.7.1",
    "pandas>=2.3.1",
    "pyopenjtalk>=0.4.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.4",
]

[tool.mypy]
# 基本設定
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.ruff]
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "T201",
    "COM812",
    "ISC001",
]
unfixable = [
    "F401",
    "F841",
]
pydocstyle.convention = "google" 

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "D",
    "S101",
    "N802",
    "ARG",
]

[tool.ruff.lint.pylint]
max-args = 6
